requirements.txt
Django==3.0.2
django-polymorphic==2.1.2
django-nested-admin==3.2.4
django-admin-sortable2==0.7.5
psycopg2-binary==2.8.4
python-dotenv==0.14.0